refactor: merge TES code into refiner application
