+++ .gitignore
@@ -21,5 +21,8 @@ node_modules/
 .coverage
 coverage.xml
 
-#Environment
-.env
+# Environment
+**/.env
+
+# Tools
+.requests
