+++ api/refiner/scripts/tests/assets/sample_ecr.json
