--- api/trigger-code-reference/.gitignore
@@ -1,1 +1,0 @@
-.env
