--- api/trigger-code-reference/assets/README.md
@@ -1,5 +1,0 @@
-# Assets
-
-## RCKMS Condition Codes
-
-This file can be downloaded from the [RCKMS site]("https://www.rckms.org/content-repository/).
