--- api/trigger-code-reference/assets/sample_get_value_sets_requests.json
@@ -1,17 +1,0 @@
-{
-  "Simple value set retrieval": {
-    "summary": "Retrieves applicable value sets for corynebacterium diphtheriae",
-    "description": "This example queries the SQLite database for value sets related to condition code 276197005; it returns all generated content (i.e. uses no section filtering).",
-    "value": {
-      "condition_code": "276197005"
-    }
-  },
-  "Filtered value set retrieval": {
-    "summary": "Retrieves and filters COVID-related value sets using a concept type filter",
-    "description": "This example queries the SQLite database for value sets related to condition code 840539006; we're interested only in lab-related services for this code, so we filter out diagnostics.",
-    "value": {
-      "condition_code": "840539006",
-      "filter_concepts": "lrtc,ostc"
-    }
-  }
-}
