--- api/trigger-code-reference/assets/sample_get_value_sets_responses.json
@@ -1,245 +1,0 @@
-{
-  "description": "Success",
-  "content": {
-    "application/json": {
-      "examples": {
-        "Simple value set retrieval": {
-          "summary": "Unfiltered value sets for condition code 276197005",
-          "value": {
-            "dxtc": [
-              {
-                "codes": [
-                  "A36.3",
-                  "A36",
-                  "A36.9",
-                  "A36.81",
-                  "A36.86",
-                  "A36.85",
-                  "A36.83",
-                  "A36.82",
-                  "A36.84",
-                  "A36.2",
-                  "A36.1",
-                  "A36.8",
-                  "A36.89",
-                  "A36.0"
-                ],
-                "system": "http://hl7.org/fhir/sid/icd-10-cm"
-              },
-              {
-                "codes": [
-                  "194945009",
-                  "15682004",
-                  "1086051000119107",
-                  "7773002",
-                  "18901009",
-                  "397428000",
-                  "397430003",
-                  "1177005009",
-                  "186347006",
-                  "715659006",
-                  "1086061000119109",
-                  "1086071000119103",
-                  "48278001",
-                  "26117009",
-                  "230596007",
-                  "129667001",
-                  "13596001",
-                  "1259059003",
-                  "3419005",
-                  "276197005",
-                  "50215002",
-                  "75589004",
-                  "789005009",
-                  "1090211000119102",
-                  "240422004"
-                ],
-                "system": "http://snomed.info/sct"
-              }
-            ],
-            "lrtc": [
-              {
-                "codes": [
-                  "16676-9",
-                  "88584-8",
-                  "14483-2",
-                  "6596-1",
-                  "14481-6",
-                  "14482-4",
-                  "567-8",
-                  "14480-8",
-                  "14479-0"
-                ],
-                "system": "http://loinc.org"
-              }
-            ],
-            "ostc": [
-              {
-                "codes": [
-                  "151000221104",
-                  "5851001",
-                  "83675005",
-                  "70876001",
-                  "13755001",
-                  "243255007",
-                  "708224008",
-                  "55123007",
-                  "443401009",
-                  "722313007",
-                  "722336003",
-                  "722790009",
-                  "722339005",
-                  "443377006",
-                  "722789000",
-                  "722346001",
-                  "722349008",
-                  "722350008"
-                ],
-                "system": "http://snomed.info/sct"
-              }
-            ],
-            "sdtc": [
-              {
-                "codes": ["772150003"],
-                "system": "http://snomed.info/sct"
-              }
-            ]
-          }
-        },
-        "Filtered value set retrieval": {
-          "summary": "Filtered, lab-related value sets for condition code 840539006",
-          "value": {
-            "lrtc": [
-              {
-                "codes": [
-                  "95423-0",
-                  "96764-6",
-                  "96763-8",
-                  "96957-6",
-                  "95521-1",
-                  "96898-2",
-                  "94510-5",
-                  "94311-8",
-                  "94312-6",
-                  "95522-9",
-                  "94760-6",
-                  "96986-5",
-                  "95409-9",
-                  "94533-7",
-                  "94756-4",
-                  "94757-2",
-                  "95425-5",
-                  "96448-6",
-                  "96958-4",
-                  "94766-3",
-                  "94316-7",
-                  "94307-6",
-                  "94308-4",
-                  "97098-8",
-                  "98132-4",
-                  "98494-8",
-                  "96899-0",
-                  "94644-2",
-                  "94511-3",
-                  "94559-2",
-                  "95824-9",
-                  "94639-2",
-                  "97104-4",
-                  "98131-6",
-                  "98493-0",
-                  "94646-7",
-                  "94645-9",
-                  "96120-1",
-                  "94534-5",
-                  "96091-4",
-                  "94314-2",
-                  "96123-5",
-                  "94745-7",
-                  "94746-5",
-                  "94819-0",
-                  "94565-9",
-                  "94759-8",
-                  "95406-5",
-                  "96797-6",
-                  "95608-6",
-                  "94500-6",
-                  "95424-8",
-                  "94845-5",
-                  "94822-4",
-                  "94660-8",
-                  "94309-2",
-                  "101289-7",
-                  "96900-6",
-                  "94642-6",
-                  "94643-4",
-                  "94640-0",
-                  "95609-4",
-                  "96765-3",
-                  "94767-1",
-                  "94641-8",
-                  "96756-2",
-                  "96757-0",
-                  "95970-0",
-                  "94509-7",
-                  "96121-9",
-                  "94758-0",
-                  "95823-1",
-                  "94765-5",
-                  "94315-9",
-                  "96122-7",
-                  "94313-4",
-                  "94310-0",
-                  "94502-2",
-                  "94647-5",
-                  "94532-9",
-                  "95209-3",
-                  "94558-4",
-                  "96119-3",
-                  "97097-0",
-                  "101928-0",
-                  "96896-6",
-                  "96895-8",
-                  "100157-7",
-                  "99314-7",
-                  "96752-1",
-                  "96751-3",
-                  "100156-9",
-                  "96741-4",
-                  "96755-4",
-                  "94764-8"
-                ],
-                "system": "http://loinc.org"
-              }
-            ],
-            "ostc": [
-              {
-                "codes": [
-                  "467951000124101",
-                  "467961000124104",
-                  "467971000124106",
-                  "840536004",
-                  "1157045004",
-                  "29301000087104",
-                  "1179416004",
-                  "1187591006",
-                  "871555000",
-                  "871559006",
-                  "871556004",
-                  "871557008",
-                  "871558003",
-                  "871560001",
-                  "871562009",
-                  "871553007",
-                  "1240411000000107",
-                  "840533007",
-                  "1240581000000104"
-                ],
-                "system": "http://snomed.info/sct"
-              }
-            ]
-          }
-        }
-      }
-    }
-  }
-}
