--- api/trigger-code-reference/pyproject.toml
@@ -1,3 +1,0 @@
-[tool.ruff]
-extend = "../../pyproject.toml"
-target-version = "py313"
