--- api/trigger-code-reference/requirements.txt
@@ -1,6 +1,0 @@
-fastapi>=0.109.1
-fhirpathpy
-pathlib
-pydantic>=2.0.0
-requests
-uvicorn
